Improve USFM to HTML conversion

+++ converters/usfm2html_converter.py
@@ -25,7 +25,7 @@ class Usfm2HtmlConverter(Converter):
 
         # Convert usfm files and copy across other files
         num_successful_books = num_failed_books = 0
-        for filename in files:
+        for filename in sorted(files):
             if filename.endswith('.usfm'):
                 base_name = os.path.basename(filename)
                 if convert_only_list and (base_name not in convert_only_list):  # see if this is a file we are to convert
@@ -35,14 +35,21 @@ class Usfm2HtmlConverter(Converter):
                 self.log.info(f"Converting Bible USFM file: {base_name} …") # Logger also issues DEBUG msg
                 # Copy just the single file to be converted into a single scratch folder
                 scratch_dir = tempfile.mkdtemp(prefix='usfm_convert_scratch_')
+                delete_scratch_dir_flag = True
                 copyfile(filename, os.path.join(scratch_dir, os.path.basename(filename)))
                 filebase = os.path.splitext(os.path.basename(filename))[0]
                 # Do the actual USFM -> HTML conversion
                 UsfmTransform.buildSingleHtml(scratch_dir, scratch_dir, filebase)
+
+                # This code seems to be cleaning up or adjusting the converted HTML file
                 html_filename = filebase + '.html'
                 with open(os.path.join(scratch_dir, html_filename), 'rt', encoding='utf-8') as html_file:
                     converted_html = html_file.read()
+                converted_html_length = len(converted_html)
+                # GlobalSettings.logger.debug(f"Usfm2HtmlConverter got converted html of length {converted_html_length:,}")
                 # GlobalSettings.logger.debug(f"Got converted html: {converted_html[:500]}{' …' if len(converted_html)>500 else ''}")
+                if '</p></p></p>' in converted_html:
+                    GlobalSettings.logger.debug(f"Usfm2HtmlConverter got multiple consecutive paragraph closures in converted {html_filename}")
                 # Now what are we doing with the converted html ???
                 template_soup = BeautifulSoup(template_html, 'html.parser')
                 template_soup.head.title.string = self.resource.upper()
@@ -66,10 +73,22 @@ class Usfm2HtmlConverter(Converter):
                     num_failed_books += 1
                 output_filepath = os.path.join(self.output_dir, html_filename)
                 #print("template_soup type is", type(template_soup)) # <class 'bs4.BeautifulSoup'>
-                write_file(output_filepath, str(template_soup))
+                template_soup_string = str(template_soup)
+                write_file(output_filepath, template_soup_string)
+                template_soup_string_length = len(template_soup_string)
+                # GlobalSettings.logger.debug(f"Usfm2HtmlConverter wrote souped-up html of length {template_soup_string_length:,} from {converted_html_length:,}")
+                if '</p></p></p>' in template_soup_string:
+                    GlobalSettings.logger.warning(f"Usfm2HtmlConverter got multiple consecutive paragraph closures in {html_filename}")
+                if template_soup_string_length < converted_html_length * 0.75: # What is the 25% that's lost ???
+                    self.log.error(f"Usfm2HtmlConverter lost converted html for {html_filename}")
+                    GlobalSettings.logger.debug(f"Usfm2HtmlConverter {html_filename} was {converted_html_length:,} now {template_soup_string_length:,}")
+                    # GlobalSettings.logger.debug(f"Usfm2HtmlConverter {html_filename} was: {converted_html}")
+                    write_file(os.path.join(scratch_dir,filebase+'.converted.html'), template_soup_string)
+                    delete_scratch_dir_flag = False
                 #print("Got converted x2 html:", str(template_soup)[:500])
-                self.log.info(f"Converted {os.path.basename(filename)} to {os.path.basename(html_filename)}.")
-                remove_tree(scratch_dir)
+                # self.log.info(f"Converted {os.path.basename(filename)} to {os.path.basename(html_filename)}.")
+                if delete_scratch_dir_flag:
+                    remove_tree(scratch_dir)
             else:
                 # Directly copy over files that are not USFM files
                 try:
@@ -80,5 +99,5 @@ class Usfm2HtmlConverter(Converter):
                     pass
         if num_failed_books and not num_successful_books:
             self.log.error(f"Conversion of all books failed!")
-        self.log.info("Finished processing Bible USFM files.")
+        self.log.info(f"Finished processing {num_successful_books} Bible USFM files.")
         return True
